Move common plan fragmenting code to common classes
Move BasePlanFragmenter to a top level class along with its associated
inner classes, and move utility methods to a utility class.

These will be used both by PlanFragmenter and IterativePlanFragmenter
